==> 1_samples.Rmd <==
---
title: "Samples of interest"
author: "Hank Sherwood and Joao Souto-Maior (code)"
date: "`r format(Sys.time(), '%B %d %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
---

```{r include=FALSE, message=FALSE, warning=FALSE}
rm(list = ls())
library(knitr)
opts_chunk$set(fig.align="center", warning=FALSE, fig.pos="H")

source("helper_heading.R")
source("helper_data.R")
source("helper_plot-var-names.R")
```

# Gender and sexuality in the data

```{r}
nrow(dat)
vars <- dat %>%
  select("gender minority",
         "sexual minority")
summary_cat(vars)
```

```{r}
freq_table(dat$"sex female")
```

```{r}
freq_table(dat$"gender minority")
```


```{r}
freq_table(dat$`sexual minority`)
```

```{r}
table(dat$"gender minority", dat$"sexual minority")
```

# Sex

Note one student was not assigned either male or female at birth.

```{r}
freq_table(dat$`sex female`)
```

```{r}
dat <- dat %>%
  mutate("sex female" = 
           factor(dat$"sex female", 
                  levels = c(1, 2), 
                  labels = c( "No", "Yes"),
                  ordered = TRUE))
freq_table(dat$`sex female`)
```

Save original data for later.

```{r}
dat_main <- dat
```

# Adjust item labels

```{r}
################## Adjust survey items
# this need to be done in more detail if we are asked to
# report the items behind all composite variables.

vars <- dat %>%
  select(starts_with("item"))
lapply(vars, summary)

items_1_4 <- dat %>%
  select(starts_with("item") & (contains("GW") | 
                                  contains("classroom"))) %>%
  names()

items_0_4 <- dat %>%
  select(starts_with("item") & (contains("involvement") | 
                                  contains("engagement") |
                                  contains("victimization"))) %>%
  names()

items_1_5 <- dat %>%
  select(starts_with("item") & (contains("belonging") | 
                                  contains("PS"))) %>%
  names()

# create factors
dat <- dat %>%
  mutate(across(.cols = all_of(c(items_1_4)),
                .fns = ~factor(x = .,
                               levels = c(1, 2, 3, 4),
                               labels = c( "1-Not true at all",
                                           "2-Somewhat false",
                                           "3-Somewhat true",
                                           "4-Very true"),
                               ordered = TRUE))) %>%
  mutate(across(.cols = all_of(c(items_0_4)),
                .fns = ~factor(x = .,
                               levels = c(0, 1, 2, 3, 4),
                               labels = c( "0-Very low",
                                           "1-Low",
                                           "2-Mid",
                                           "3-High",
                                           "4-Very high"),
                               ordered = TRUE))) %>%
  mutate(across(.cols = all_of(c(items_1_5)),
                .fns = ~factor(x = .,
                               levels = c(1, 2, 3, 4, 5),
                               labels = c( "1-Very low",
                                           "2-Low",
                                           "3-Mid",
                                           "4-High",
                                           "5-Very high"),
                               ordered = TRUE)))
  
# check
vars <- dat %>%
  select(starts_with("item"))
#lapply(vars, summary)

################### var labels (if needed...)
# e.g.
# data = expss::apply_labels(dat, parental.ed = "Parental education")
```

# Gender 

Only keep those who with no Nas in the outness questions.

## Describe outness items

```{r }
dat_sex <- dat %>%
  filter(dat$"sexual minority" == "Yes") %>%
  select(-starts_with("GenOut"))
```

```{r}
dat_item_sex <- dat_sex %>%
  select(starts_with("SexOut"))

tab <- summary_cat(dat_item_sex)
tab
```

```{r}
dat_gen <- dat %>%
  filter(dat$"gender minority" == "Yes") %>%
  select(-starts_with("SexOut"))
```

```{r}
dat_item_gen <- dat_gen %>%
  select(starts_with("GenOut"))

tab <- summary_cat(dat_item_gen)
tab
```

# Define separate samples

## Gender minority data (and most frequent categories)

```{r}
nrow(dat_gen)
dat_gen <- dat_gen %>%
  drop_na(starts_with("GenOut"), 
          -contains("siblings"))
nrow(dat_gen)
```

```{r}
dat <- dat_gen
source("helper_code-var-names.R")
dat_gen <- dat
write_rds(dat_gen, "../data/dat_gen.Rdata")
```

```{r}
freq_table(dat_gen$sex.female)
```

```{r}
freq_table(dat_gen$nonbinary)
freq_table(dat_gen$genderqueer)
freq_table(dat_gen$genderfluid)
freq_table(dat_gen$agender)
freq_table(dat_gen$genderspecify)
```

## Sexual minority data (and most frequent categories)


```{r}
nrow(dat_sex)
dat_sex <- dat_sex %>%
  drop_na(starts_with("SexOut"),
          -contains("siblings"))
nrow(dat_sex)
```

```{r}
dat <- dat_sex
source("helper_code-var-names.R")
dat_sex <- dat
write_rds(dat_sex, "../data/dat_sex.Rdata")
```

```{r}
freq_table(dat_sex$sex.female)
```

```{r}
freq_table(dat_sex$gaylesbian)
freq_table(dat_sex$bisexual)
freq_table(dat_sex$questioning)
freq_table(dat_sex$pansexual)
freq_table(dat_sex$asexual)
freq_table(dat_sex$queer)
freq_table(dat_sex$sospecify)
```

# Cronbach alpha measures for selected variables

Main data.

```{r}
dat <- dat_main %>%
  select(-starts_with("SexOut"),
         -starts_with("GenOut"),
         -`sex female`)
  
source("helper_code-var-names.R")

dat_sex <- dat_sex %>%
  select(ID_student) %>%
  mutate(sample.sex = 1)
dat_gen <- dat_gen %>%
  select(ID_student)%>%
  mutate(sample.gen = 1)

dat <- left_join(dat, dat_sex, by = c("ID_student" = "ID_student")) 
dat <- left_join(dat, dat_gen, by = c("ID_student" = "ID_student")) 

dat <- dat %>%
  filter(sample.gen == 1 | sample.sex == 1)
```

```{r}
#cfa_model  <- ' emotional.engagement.check  = ~ item.1.classroom + #item.2.classroom + #item.3.classroom + item.4.classroom + item.5.classroom'
#cfa <- cfa(cfa_model, data = dat) 
#summary(cfa, fit.measures=TRUE)
```

## Behavioral engagement 

```{r}
factor <- dat %>%
  select(item.1.classroom,
         item.2.classroom,
         item.3.classroom,
         item.4.classroom,
         item.5.classroom) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```

## Emotional engagement

```{r}
factor <- dat %>%
  select(item.6.classroom,
         item.7.classroom,
         item.8.classroom,
         item.9.classroom,
         item.10.classroom) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```
## Behavioral disafection

```{r}
factor <- dat %>%
  select(item.11.classroom,
         item.12.classroom,
         item.13.classroom,
         item.14.classroom) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```

## Emotional disafection

```{r}
factor <- dat %>%
  select(item.15.classroom,
         item.16.classroom,
         item.17.classroom,
         item.18.classroom,
         item.19.classroom) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```

## Global worth

```{r}
factor <- dat %>%
  select(item.1.GW,
         item.2.GW,
         item.3.GW,
         item.4.GW,
         item.5.GW) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```

## GSA involvement

```{r}
factor <- dat %>%
  select(item.1.involvement,
         item.2.involvement,
         item.3.involvement,
         item.4.involvement,
         item.5.involvement,
         item.6.involvement,
         item.7.involvement) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```

## Peer sensitivity

```{r}
factor <- dat %>%
  select(item.1.PS,
         item.2.PS,
         item.3.PS,
         item.4.PS,
         item.5.PS,
         item.6.PS,
         item.7.PS) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```

## GSA engagement

```{r}
factor <- dat %>%
  select(item.1.engagement,
         item.2.engagement,
         item.3.engagement,
         item.4.engagement,
         item.5.engagement) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```

## Victimization

```{r}
factor <- dat %>%
  select(item.1.victimization,
         item.2.victimization,
         item.3.victimization,
         item.4.victimization,
         item.5.victimization ) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```

## Victimization.LGBT

```{r}
factor <- dat %>%
  select(item.1.victimization.LGBT,
         item.2.victimization.LGBT,
         item.3.victimization.LGBT,
         item.4.victimization.LGBT,
         item.5.victimization.LGBT) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```

## GSA belonging

```{r}
factor <- dat %>%
  select(item.1.belonging,
         item.2.belonging,
         item.3.belonging,
         item.4.belonging,
         item.5.belonging,
         item.6.belonging,
         item.7.belonging,
         item.8.belonging,
         item.9.belonging,
         item.10.belonging) %>%
  as.data.frame()

a = psych::alpha(factor)
a = a[[1]]
a$raw_alpha
```
