==> 3b_ols-results-gender.Rmd <==
---
title: "Results: gender minority sample"
author: "Hank Sherwood and Joao Souto-Maior (code)"
date: "`r format(Sys.time(), '%B %d %Y')`"
output:
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
---

```{r include=FALSE, message=FALSE, warning=FALSE}
rm(list = ls())
library(knitr)
opts_chunk$set(fig.align="center", warning=FALSE, fig.pos="H")
```

# Set up

```{r message=FALSE, warning=FALSE}
source("helper_heading.R")
dat <- read_rds("../data/dat_gen_2.Rdata")
source("helper_var-groups.R")
```

# Data

```{r}
dat <- dat %>%
  select(-starts_with("item"),
         -starts_with("GenOut"),
         -sexual.minority, 
         -gender.minority) %>%
  mutate(outness.about.gender.identity = outness.about.gender.identity + 4)
glimpse(dat)
```

# Missing data

```{r}
pct_missing_variable(dat)
```

## Mutiple imputation

```{r}
set.seed(10)
imp <- mice(dat, maxit=0)
predM <- imp$predictorMatrix
meth <- imp$method
```

Define imputation model for given variables.

```{r}
dummy <- dat %>%
  select(sex.female,
         race.non.white) %>%
  names()
cat <- dat %>%
  select(parental.ed) %>%
  names()

#meth[ordered_cat] <- "polr"
meth[dummy] <- "logreg"
meth[cat] <- "polyreg"
meth
```

Imputation leading to 5 datasets.

```{r}
m_imputation <- mice(dat, maxit = 5, 
                     predictorMatrix = predM, 
                     method = meth, 
                     print =  FALSE)
```

Imputed datasets.

```{r}
dat1 <- mice::complete(m_imputation, 1)
dat2 <- mice::complete(m_imputation, 2)
dat3 <- mice::complete(m_imputation, 3)
dat4 <- mice::complete(m_imputation, 4)
dat5 <- mice::complete(m_imputation, 5)
```

Illustration.

```{r}
freq_table(dat1$parental.ed)
freq_table(dat2$parental.ed)
```

# Check nested structure of data

```{r}
t(sapply(dat[, c("ID_school", "ID_student")], summary_levels)) 
```

```{r message=FALSE, warning=F}
source("helper_check-nested-structure.R")
check_nested_structure(emotional.engagement, 
                       outness.about.gender.identity, 
                       ID_school,
                       ID_student,
                       dat1)
```

School-level correlations do not seem very strong. That said, correlation within groups, even when small, are problematic for traditional OLS models for two main reasons: (1) the independence of errors assumption is unlikely to hold, which bias the standard errors of coefficients and lead to unreliable results concerning the statistical significance of coefficients; (2) if we do not account for an existing within-group correlation, we might introduce an issue of omitted variable bias, which might hinder coefficients themselves to be unreliable.

Below I provide two options to account for the within-school correlation presented in this nested data: (a) robust standard errors in an OLS model (White, 1980); (b) school-specific effects in multilevel regression model. Ideally, option (b) is preferred because it fixes both problems (1) and (2). That said, as I demonstrate below, a multilevel model fails to converge because the structure of the model is not supported by the underlying data. This is likely due to the limit number of observations in many schools in this sample. Given this limitation, I consider solution (b) to minimize the the issue of within-group correlation. 

# Models

```{r message=FALSE, warning=FALSE}
source("helper_models.R")
```

## Emotional engagement

Run models.

```{r}
list = list(dat1, dat2, dat3, dat4, dat5)

m1 <- ols_model1("emotional.engagement", "outness.about.gender.identity", list)
m2 <- ols_model2("emotional.engagement", "outness.about.gender.identity", list)
m3 <- ols_model3("emotional.engagement", "outness.about.gender.identity", list)

m_mlm_1 <- ols_model1("emotional.engagement", "outness.about.gender.identity", list)
m_mlm_2 <- ols_model2("emotional.engagement", "outness.about.gender.identity", list)
m_mlm_3 <- ols_model3("emotional.engagement", "outness.about.gender.identity", list)
```

### Pooled results for OLS with robust SEs (pooled across imputed datasets)

```{r}
r1 <- pool(m1)
r2 <- pool(m2)
r3 <- pool(m3)
r_emotional = r3

screenreg(list(r1, r2, r3))

# Omit controls
screenreg(list(r1, r2, r3),
          omit.coef = c("race.non.white|sex.female|parental.ed|grade|age$"))
```

### Pooled results for MLM (does not converge)

```{r}
r1 <- pool(m_mlm_1)
r2 <- pool(m_mlm_2)
r3 <- pool(m_mlm_3)

screenreg(list(r1, r2, r3))

# Omit controls
screenreg(list(r1, r2, r3),
          omit.coef = c("race.non.white|sex.female|parental.ed|grade|age$"))

```

## Behavioral engagement

Run models.

```{r}
list = list(dat1, dat2, dat3, dat4, dat5)

m1 <- ols_model1("behavioral.engagement", "outness.about.gender.identity", list)
m2 <- ols_model2("behavioral.engagement", "outness.about.gender.identity", list)
m3 <- ols_model3("behavioral.engagement", "outness.about.gender.identity", list)

m_mlm_1 <- ols_model1("behavioral.engagement", "outness.about.gender.identity", list)
m_mlm_2 <- ols_model2("behavioral.engagement", "outness.about.gender.identity", list)
m_mlm_3 <- ols_model3("behavioral.engagement", "outness.about.gender.identity", list)
```

### Pooled results for OLS with robust SEs (pooled across imputed datasets)

```{r}
r1 <- pool(m1)
r2 <- pool(m2)
r3 <- pool(m3)
r_behavioral = r3

screenreg(list(r1, r2, r3))

# Omit controls
screenreg(list(r1, r2, r3),
          omit.coef = c("race.non.white|sex.female|parental.ed|grade|age$"))
```

### Pooled results for MLM (does not converge)

```{r}
r1 <- pool(m_mlm_1)
r2 <- pool(m_mlm_2)
r3 <- pool(m_mlm_3)

screenreg(list(r1, r2, r3))

# Omit controls
screenreg(list(r1, r2, r3),
          omit.coef = c("race.non.white|sex.female|parental.ed|grade|age$"))

```

## Plot interaction effect (Behavioral engagement)

Note: simulations will be run for a student with variable values which capture average values from the data. For categorical variables, the student is defined by the most frequent category. 

Just one quick check on the reference category of parental background.

```{r}
freq_table(dat$parental.ed)
```

Get the correct data.

```{r}
vars <- c("outness.about.gender.identity", 
          "GSA.engagement",
          vars_controls,
          var_mechanisms)

model_dat <- dat1 %>%
  select(all_of(vars))

#transform factors into numeric to average the data for simulations
factor_vars <- model_dat %>%
  select(!is.numeric) %>%
  names()
# factor vars in the data
factor_vars

#convert factor to numeric (make sure we ask for integers)
model_dat <- model_dat %>%
  mutate(across(.cols = all_of(c(factor_vars)),
                .fns = ~as.integer(.)))
glimpse(model_dat)

# calculate mean of data
mean_dat1 <- mean_of_data(model_dat)

# plug in the three factor vars with most frequent values in the data
# check the descriptive stats about the data to see values
mean_dat1 <- mean_dat1 %>%
  mutate(race.non.white = 0,
         sex.female = 1,
         parental.ed = 1)

###########################################
# do the same of a second dataset (will be useful later)

# calculate mean of data
mean_dat2 <- mean_of_data(model_dat)

# plug in the three factor vars with most frequent values in the data
# check the descriptive stats about the data to see values
mean_dat2 <- mean_dat2 %>%
  mutate(race.non.white = 0,
         sex.female = 1,
         parental.ed = 1)
```

```{r}
# check the distribution before deciding on how many
# values to estimate
hist(dat1$outness.about.gender.identity)
hist(dat1$GSA.engagement)
```

```{r}
coeffs <- as_tibble(r_behavioral$pooled) %>%
  select(term, estimate) %>%
  as_tibble(t(.))

# print coeffs to see the position of coeffs
coeffs
```

```{r}
# simulation for the 25th percentile

# add x and y values to the average data
# 81 is the number of total rows in the data
# which is defined by the number of simulated y values we want
mean_dat1 <- mean_dat1 %>% 
  slice(rep(1:n(), each = 81)) %>%
  mutate(outness.about.gender.identity = seq(0,8, by = 0.1),
         GSA.engagement = quantile(dat1$GSA.engagement, probs = c(0.25))) 

mean_dat1 <- mean_dat1 %>%
  mutate(value = coeffs$estimate[[1]] + #intercept
           age * coeffs$estimate[[13]] + 
           global.worth *coeffs$estimate[[14]]+ 
           grade *coeffs$estimate[[12]]+ 
           GSA.belonging *coeffs$estimate[[19]]+ 
           GSA.engagement *coeffs$estimate[[13]]+ 
           LGBT.victimization *coeffs$estimate[[16]]+ 
           outness.about.gender.identity *coeffs$estimate[[2]]+ 
           parental.ed *coeffs$estimate[[3]]+ #college degree
           peer.sensitivity *coeffs$estimate[[18]]+ 
           race.non.white *coeffs$estimate[[4]]+ 
           sex.female *coeffs$estimate[[5]]+ 
           social.support *coeffs$estimate[[15]]+ 
           victimization *coeffs$estimate[[17]] +
           outness.about.gender.identity*GSA.engagement*coeffs$estimate[[20]])

###############################################
# simulation for the 75th percentile

mean_dat2 <- mean_dat2 %>% 
  slice(rep(1:n(), each = 81)) %>%
  mutate(outness.about.gender.identity = seq(0,8, by = 0.1),
         GSA.engagement = quantile(dat1$GSA.engagement, probs = c(0.75))) 

mean_dat2 <- mean_dat2 %>%
  mutate(value = coeffs$estimate[[1]] + #intercept
           age * coeffs$estimate[[13]] + 
           global.worth *coeffs$estimate[[14]]+ 
           grade *coeffs$estimate[[12]]+ 
           GSA.belonging *coeffs$estimate[[19]]+ 
           GSA.engagement *coeffs$estimate[[13]]+ 
           LGBT.victimization *coeffs$estimate[[16]]+ 
           outness.about.gender.identity *coeffs$estimate[[2]]+ 
           parental.ed *coeffs$estimate[[3]]+ #college degree
           peer.sensitivity *coeffs$estimate[[18]]+ 
           race.non.white *coeffs$estimate[[4]]+ 
           sex.female *coeffs$estimate[[5]]+ 
           social.support *coeffs$estimate[[15]]+ 
           victimization *coeffs$estimate[[17]] +
           outness.about.gender.identity*GSA.engagement*coeffs$estimate[[20]])
```

```{r}
p <- 
  ggplot(mean_dat1,
         aes(x = outness.about.gender.identity, 
                 y = value)) +
  geom_point(alpha = 0.8,
             shape = 19,
             size = 2,
             color = "red") +
  geom_point(data = mean_dat2,
             aes(x = outness.about.gender.identity, 
                 y = value),
             alpha = 0.8,
             shape = 19,
             size = 2,
             color = "blue") +
  ylim(c(0,2.5)) +
  labs(x = "Level of outness about gender identity",
       y = "Behavioral engagement") + 
  geom_text(mapping = aes(x = 2, y = 0, label = paste("GSA engagement = 25th percentile in the data")),
            size = 4, 
            angle = 0,
            alpha = 0.1,
            color = "red") +
  geom_text(mapping = aes(x = 2, y = 0.2, label = paste("GSA engagement = 75th percentile in the data")),
            size = 4,
            angle = 0,
            alpha = 0.1,
            color = "blue") +
  theme_classic()
p

ggsave("figures/behavioral_eng_gen_interaction.svg")
```

## Plot interaction effect (emotional)

```{r}
vars <- c("outness.about.gender.identity", 
          "GSA.engagement",
          vars_controls,
          var_mechanisms)

model_dat <- dat1 %>%
  select(all_of(vars))

#transform factors into numeric to average the data for simulations
factor_vars <- model_dat %>%
  select(!is.numeric) %>%
  names()
# factor vars in the data
factor_vars

#convert factor to numeric (make sure we ask for integers)
model_dat <- model_dat %>%
  mutate(across(.cols = all_of(c(factor_vars)),
                .fns = ~as.integer(.)))
glimpse(model_dat)

# calculate mean of data
mean_dat1 <- mean_of_data(model_dat)

# plug in the three factor vars with most frequent values in the data
# check the descriptive stats about the data to see values
mean_dat1 <- mean_dat1 %>%
  mutate(race.non.white = 0,
         sex.female = 1,
         parental.ed = 1)

###########################################
# do the same of a second dataset (will be useful later)

# calculate mean of data
mean_dat2 <- mean_of_data(model_dat)

# plug in the three factor vars with most frequent values in the data
# check the descriptive stats about the data to see values
mean_dat2 <- mean_dat2 %>%
  mutate(race.non.white = 0,
         sex.female = 1,
         parental.ed = 1)
```

```{r}
coeffs <- as_tibble(r_emotional$pooled) %>%
  select(term, estimate) %>%
  as_tibble(t(.))

# print coeffs to see the position of coeffs
coeffs
```

```{r}
# simulation for the 25th percentile

# add x and y values to the average data
# 81 is the number of total rows in the data
# which is defined by the number of simulated y values we want
mean_dat1 <- mean_dat1 %>% 
  slice(rep(1:n(), each = 81)) %>%
  mutate(outness.about.gender.identity = seq(0,8, by = 0.1),
         GSA.engagement = quantile(dat1$GSA.engagement, probs = c(0.25)))  

mean_dat1 <- mean_dat1 %>%
  mutate(value = coeffs$estimate[[1]] + #intercept
           age * coeffs$estimate[[13]] + 
           global.worth *coeffs$estimate[[14]]+ 
           grade *coeffs$estimate[[12]]+ 
           GSA.belonging *coeffs$estimate[[19]]+ 
           GSA.engagement *coeffs$estimate[[13]]+ 
           LGBT.victimization *coeffs$estimate[[16]]+ 
           outness.about.gender.identity *coeffs$estimate[[2]]+ 
           parental.ed *coeffs$estimate[[3]]+ #college degree
           peer.sensitivity *coeffs$estimate[[18]]+ 
           race.non.white *coeffs$estimate[[4]]+ 
           sex.female *coeffs$estimate[[5]]+ 
           social.support *coeffs$estimate[[15]]+ 
           victimization *coeffs$estimate[[17]] +
           outness.about.gender.identity*GSA.engagement*coeffs$estimate[[20]])

###############################################
# simulation for the 75th percentile

mean_dat2 <- mean_dat2 %>% 
  slice(rep(1:n(), each = 81)) %>%
  mutate(outness.about.gender.identity = seq(0,8, by = 0.1),
         GSA.engagement = quantile(dat1$GSA.engagement, probs = c(0.75))) 

mean_dat2 <- mean_dat2 %>%
  mutate(value = coeffs$estimate[[1]] + #intercept
           age * coeffs$estimate[[13]] + 
           global.worth *coeffs$estimate[[14]]+ 
           grade *coeffs$estimate[[12]]+ 
           GSA.belonging *coeffs$estimate[[19]]+ 
           GSA.engagement *coeffs$estimate[[13]]+ 
           LGBT.victimization *coeffs$estimate[[16]]+ 
           outness.about.gender.identity *coeffs$estimate[[2]]+ 
           parental.ed *coeffs$estimate[[3]]+ #college degree
           peer.sensitivity *coeffs$estimate[[18]]+ 
           race.non.white *coeffs$estimate[[4]]+ 
           sex.female *coeffs$estimate[[5]]+ 
           social.support *coeffs$estimate[[15]]+ 
           victimization *coeffs$estimate[[17]] +
           outness.about.gender.identity*GSA.engagement*coeffs$estimate[[20]])
```

```{r}
p <- 
  ggplot(mean_dat1,
         aes(x = outness.about.gender.identity, 
                 y = value)) +
  geom_point(alpha = 0.8,
             shape = 19,
             size = 2,
             color = "red") +
  geom_point(data = mean_dat2,
             aes(x = outness.about.gender.identity, 
                 y = value),
             alpha = 0.8,
             shape = 19,
             size = 2,
             color = "blue") +
  ylim(c(0,2.5)) +
  labs(x = "Level of outness about gender identity",
       y = "Emotional engagement") + 
  geom_text(mapping = aes(x = 2, y = 0, label = paste("GSA engagement = 25th percentile in the data")),
            size = 4, 
            angle = 0,
            alpha = 0.1,
            color = "red") +
  geom_text(mapping = aes(x = 2, y = 0.2, label = paste("GSA engagement = 75th percentile in the data")),
            size = 4,
            angle = 0,
            alpha = 0.1,
            color = "blue") +
  theme_classic()
p

ggsave("figures/emotional_eng_gen_interaction.svg")
```
